Safari-609.1.11.1/LayoutTests/svg/dom/feFlood-no-in1.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<script>
description("This test checks that an SVGFEFloodElement object has no in1 property.");

var feFlood = document.createElementNS("http://www.w3.org/2000/svg", "feFlood");
shouldBe("feFlood.in1", "undefined");

successfullyParsed = true;
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
